Add files via upload
DataModule is just trying shit out. i delete most stuff i try there anyway so ignore it.

The dataset now used in DecForest is all african countries, acces to electricity & real LCU GDP (apporved by samira)
here are some NaNs in the early years but the idea is to get a correlation between acces to electricity and this GDP figure.

TODO: how to train for a label that is a number i.e. our gdp